climbing: add underline for extend+none in FloatingMenu
